requirements.txt
aiogram==3.13.0
feedparser
aiosqlite
apscheduler
python-dotenv
